Fixed copilot comments

+++ src/backend/pagos/app/models/pago.py
@@ -1,12 +1,11 @@
-from sqlalchemy import Column, String, Float
+from sqlalchemy import Column, String, Numeric
 from sqlalchemy.dialects.postgresql import UUID
 from travelhub_common.models import BaseModel
-from datetime import datetime, timezone
 
 class Pago(BaseModel):
     __tablename__ = "pagos"
 
-    monto = Column(Float, nullable=False, default=0)
+    monto = Column(Numeric(16, 2), nullable=False, default=0)
     medio_de_pago = Column(String(100), nullable=False)
     estado = Column(String(50), nullable=False)
